move constructor static methods to classes
