+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/PowerplayLift.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.robot;
 
+import com.arcrobotics.ftclib.hardware.motors.MotorEx;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.teamcode.control.controller.PIDFController;
@@ -15,12 +16,29 @@ public class PowerplayLift extends ProfiledLift {
         FLOOR, TWO, THREE, FOUR, FIVE, LOW, MED, TALL
     }
 
+    public static MotorEx getLiftMotor(HardwareMap hw, String name) {
+        return new MotorEx(hw, name, 145.1, 1150);
+    }
+
+    public static MotorEx[] getLiftMotors(HardwareMap hw) {
+
+        MotorEx liftMotor1 = getLiftMotor(hw, "lift motor 1");
+        MotorEx liftMotor2 = getLiftMotor(hw, "lift motor 2");
+        MotorEx liftMotor3 = getLiftMotor(hw, "lift motor 3");
+
+        liftMotor2.setInverted(false);
+        liftMotor1.setInverted(true);
+        liftMotor3.setInverted(true);
+
+        return new MotorEx[]{liftMotor2, liftMotor1, liftMotor3};
+    }
+
     /**
      * Initialize fields <p>
      * Use {@link #updateConstants} to update constants
      */
     public PowerplayLift(HardwareMap hw) {
-        super(PowerplayScorer.getLiftMotors(hw), hw.voltageSensor.iterator().next(), new PIDFController(), new FIRLowPassFilter(), new FIRLowPassFilter());
+        super(getLiftMotors(hw), hw.voltageSensor.iterator().next(), new PIDFController(), new FIRLowPassFilter(), new FIRLowPassFilter());
         updateConstants();
     }
