+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/PowerplayPassthrough.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.robot;
 
+import com.arcrobotics.ftclib.hardware.SimpleServo;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.teamcode.systems.ProfiledClawArm;
@@ -7,12 +8,16 @@ import org.firstinspires.ftc.teamcode.systems.SimpleClaw;
 
 public class PowerplayPassthrough extends ProfiledClawArm {
 
+    public static SimpleServo getAxon(HardwareMap hw, String name) {
+        return new SimpleServo(hw, name, 0, 355);
+    }
+
     public PowerplayPassthrough(HardwareMap hw) {
         super(
-                new SimpleClaw(PowerplayScorer.getAxon(hw, "claw right"), RobotConfig.ANGLE_CLAW_OPEN, RobotConfig.ANGLE_CLAW_CLOSED),
-                PowerplayScorer.getReversedServo(PowerplayScorer.getAxon(hw, "claw pivot")),
-                PowerplayScorer.getAxon(hw, "passthrough 1"),
-                PowerplayScorer.getReversedServo(PowerplayScorer.getAxon(hw, "passthrough 2"))
+                new SimpleClaw(getAxon(hw, "claw right"), RobotConfig.ANGLE_CLAW_OPEN, RobotConfig.ANGLE_CLAW_CLOSED),
+                PowerplayScorer.getReversedServo(getAxon(hw, "claw pivot")),
+                getAxon(hw, "passthrough 1"),
+                PowerplayScorer.getReversedServo(getAxon(hw, "passthrough 2"))
         );
         updateConstants();
         currentAngle = ANGLE_FRONT;
